requirements.txt
Django==1.9.2
djangorestframework
django-rest-auth[with_social]